made changes as per feedback on PR
