+++ portfolio-project-react/src/components/Details/Details.jsx
@@ -1,14 +1,47 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import About from "../About/About";
 import WorkExperience from "../WorkExperience/WorkExperience";
 import styles from "./Details.module.css";
 
 const Details = ({ isEditModeEnabled }) => {
-  const [totalWorkExperience, setTotalWorkExperience] = useState([0, 0]);
+  const [totalWorkExperience, setTotalWorkExperience] = useState({
+    years: 0,
+    months: 0,
+  });
+  const [workExperienceList, setWorkExperienceList] = useState([]);
 
-  const UpdateTotalExperience = (totalExperienceNewValue) => {
-    setTotalWorkExperience(totalExperienceNewValue);
-  };
+  useEffect(() => {
+    const calculateTotalExp = () => {
+      const arrayNoOfDays = workExperienceList.map((item) => {
+        return calculateNoOfdays(item.startDate, item.endDate);
+      });
+
+      const totalExperienceInDays = arrayNoOfDays.reduce(
+        (previousValue, currentValue, index) => previousValue + currentValue,
+        0
+      );
+
+      const years = Math.floor(totalExperienceInDays / 365);
+
+      const remainingDays = totalExperienceInDays % 365;
+
+      const months = Math.floor(remainingDays / 30);
+
+      setTotalWorkExperience({ years: years, months: months });
+    };
+    
+    const calculateNoOfdays = (startDate, endDate) => {
+      if (endDate === "") {
+        endDate = new Date().getTime();
+      }
+      return (
+        (new Date(endDate).getTime() - new Date(startDate).getTime()) /
+        (1000 * 3600 * 24)
+      );
+    };
+
+    calculateTotalExp();
+  }, [workExperienceList]);
 
   return (
     <div className={styles.detailsSection}>
@@ -18,7 +51,8 @@ const Details = ({ isEditModeEnabled }) => {
       />
       <WorkExperience
         isEditModeEnabled={isEditModeEnabled}
-        onUpdateTotalExperience={UpdateTotalExperience}
+        workExperienceList={workExperienceList}
+        onUpdateWorkExperienceList={setWorkExperienceList}
       />
     </div>
   );
